Better support for non-terminals and monochrome terminals

+++ rootfs/etc/profile.d/_07-term-mode.sh
@@ -22,18 +22,26 @@
 function _is_term_dark_mode() {
 	local x fg_rgb bg_rgb fg_lum bg_lum
 
-	# Extract the RGB values of the foreground and background colors via OSC 10 and 11.
-	# Redirect output to `/dev/tty` in case we are in a subshell where output is a pipe,
-	# because this output has to go directly to the terminal.
-	stty -echo
-	echo -ne '\e]10;?\a\e]11;?\a' >/dev/tty
-	IFS=: read -t 0.1 -d $'\a' x fg_rgb
-	IFS=: read -t 0.1 -d $'\a' x bg_rgb
-	stty echo
+	# Do not try to auto-detect if we are not in a terminal
+	# or if termcap does not think we are in a color terminal
+	if tty -s && [[ -n "$(tput setaf 1 2>/dev/null)" ]]; then
+		# Extract the RGB values of the foreground and background colors via OSC 10 and 11.
+		# Redirect output to `/dev/tty` in case we are in a subshell where output is a pipe,
+		# because this output has to go directly to the terminal.
+		stty -echo
+		echo -ne '\e]10;?\a\e]11;?\a' >/dev/tty
+		IFS=: read -t 0.1 -d $'\a' x fg_rgb
+		IFS=: read -t 0.1 -d $'\a' x bg_rgb
+		stty echo
+	else
+		if [[ $GEODESIC_TRACE =~ "terminal" ]]; then
+			echo "* TERMINAL TRACE: ${FUNCNAME[0]} called, but not running in a color terminal." >&2
+		fi
+	fi
 
 	if [[ -z $fg_rgb ]] || [[ -z $bg_rgb ]]; then
-		if [[ $GEODESIC_TRACE =~ "terminal" ]]; then
-			echo $(tput setaf 1)* TRACE: "Terminal did not respond to OSC 10 and 11 queries.$(tput sgr0)" >&2
+		if [[ $GEODESIC_TRACE =~ "terminal" ]] && tty -s; then
+			echo "$(tput setaf 1)* TERMINAL TRACE: Terminal did not respond to OSC 10 and 11 queries.$(tput sgr0)" >&2
 		fi
 		# If we cannot determine the color scheme, we assume light mode for historical reasons.
 		if [[ "$*" =~ -b ]] || [[ "$*" =~ -m ]]; then
@@ -116,7 +124,7 @@ function _srgb_to_luminance() {
 	if [[ -z $color ]]; then
 		if [[ $GEODESIC_TRACE =~ "terminal" ]]; then
 			# Use tput and sgr0 here because this is early in the startup sequence and trace logging
-			echo "$(tput setaf 1)* TRACE: ${FUNCNAME[0]} called with empty or no argument.$(tput sgr0)" >&2
+			echo "$(tput setaf 1)* TERMINAL TRACE: ${FUNCNAME[0]} called with empty or no argument.$(tput sgr0)" >&2
 		fi
 		echo "0"
 		return
